refactor(cli): split milestone dossier renderer
